chore: generated with latest sdk-generation

+++ pkg/models/shared/error.go
@@ -2,5 +2,5 @@ package shared
 
 type Error struct {
 	Message    string `json:"message"`
-	StatusCode int64  `json:"status_code"`
+	StatusCode int32  `json:"status_code"`
 }
